Added triple graphics buffering
Some cleanings:
modified trunk/ctp2_code/ai/strategy/scheduler/scheduler.cpp
modified trunk/ctp2_code/gfx/gfx_utils/ColorSet.cpp
modified trunk/ctp2_code/gs/gameobj/Diffcly.cpp
modified trunk/ctp2_code/robot/pathing/CityAstar.cpp
modified trunk/ctp2_code/ui/aui_ctp2/c3blitter.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directblitter.cpp
modified trunk/ctp2_code/ui/aui_utils/primitives.cpp
modified trunk/ctp2_code/ui/aui_utils/primitives.h

Enhanced debug memory reporting:
modified trunk/ctp2_code/ctp/debugtools/debugmemory.cpp

Added triple buffering, the objects are not drawn on the primate surface anymore, the screen you seen but on a back buffer, this makes the screen more smooth and removes a screen flickering artifact, which caused the screen to blink black for a second:
modified trunk/ctp2_code/ctp/civ3_main.cpp
modified trunk/ctp2_code/ctp/civapp.cpp
modified trunk/ctp2_code/ui/aui_common/aui_bitmapfont.cpp
modified trunk/ctp2_code/ui/aui_common/aui_dropdown.cpp
modified trunk/ctp2_code/ui/aui_common/aui_image.cpp
modified trunk/ctp2_code/ui/aui_common/aui_mouse.cpp
modified trunk/ctp2_code/ui/aui_common/aui_movie.cpp
modified trunk/ctp2_code/ui/aui_common/aui_moviebutton.cpp
modified trunk/ctp2_code/ui/aui_common/aui_surface.cpp
modified trunk/ctp2_code/ui/aui_common/aui_surface.h
modified trunk/ctp2_code/ui/aui_common/aui_ui.cpp
modified trunk/ctp2_code/ui/aui_common/aui_ui.h
modified trunk/ctp2_code/ui/aui_directx/aui_directmouse.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directmovie.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directsurface.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directsurface.h
modified trunk/ctp2_code/ui/aui_directx/aui_directui.cpp
modified trunk/ctp2_code/ui/aui_directx/aui_directui.h
modified trunk/ctp2_code/ui/interface/splash.cpp


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1038 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ai/strategy/scheduler/scheduler.cpp
@@ -301,7 +301,7 @@ void Scheduler::Cleanup()
 	m_goals_of_type.clear();
 	m_goals.clear();
 
-	for(int i = 0; i < m_generic_goals.size(); ++i)
+	for(size_t i = 0; i < m_generic_goals.size(); ++i)
 	{
 		delete m_generic_goals[i];
 	}
@@ -348,7 +348,7 @@ void Scheduler::SetPlayerId(const PLAYER_INDEX &player_index)
 {
 	m_playerId = player_index;
 
-	for(int i = 0; i < m_generic_goals.size(); ++i)
+	for(size_t i = 0; i < m_generic_goals.size(); ++i)
 	{
 		m_generic_goals[i]->Set_Player_Index(player_index);
 	}
